docs(core): write docs and examples
