+++ packages/core/data-model/mod.ts
@@ -1,7 +1,73 @@
 /**
- * Iroha data model.
+ * Iroha data model - types and codecs.
  *
- * TODO: describe what are these types, how to use them. Describe enums. Describe how to use codecs.
+ * It is based on the `schema.json` generated by Iroha. The schema itself is also available, see the `@iroha/core/data-model/schema` and `@iroha/core/data-model/schema-json` modules.
+ *
+ * ## Encoding
+ *
+ * Each data model type has a codec to encode/decode it from [SCALE representation](https://docs.substrate.io/reference/scale-codec/)
+ * (a simple binary format).
+ *
+ * For example, here is how to encode and decode a {@linkcode BlockStatus}:
+ *
+ * ```ts
+ * import { getCodec } from '@iroha/core'
+ * import * as types from '@iroha/core/data-model'
+ * import { assertEquals } from '@std/assert/equals'
+ * import { encodeHex } from '@std/encoding/hex'
+ *
+ * const value = types.BlockStatus.Rejected.ConsensusBlockRejection
+ *
+ * const encoded: Uint8Array = getCodec(types.BlockStatus).encode(value)
+ * assertEquals(encodeHex(encoded), '0200')
+ *
+ * const decoded = getCodec(types.BlockStatus).decode(encoded)
+ * assertEquals(decoded, value)
+ * ```
+ *
+ * ## Enumerations
+ *
+ * The data model contains many enumeration (i.e. discriminated unions), represented either as {@linkcode Variant}
+ * (`kind` + `value`) or {@linkcode VariantUnit} (just `kind`). An example of it is {@linkcode AssetType}:
+ *
+ * ```ts
+ * import * as types from '@iroha/core/data-model'
+ *
+ * const store: types.AssetType = { kind: 'Store' }
+ * const numeric: types.AssetType = { kind: 'Numeric', value: { scale: 5 } }
+ * ```
+ *
+ * Alternatively, enums could be constructed with pre-generated constructors, which makes it less verbose:
+ *
+ * ```ts
+ * import * as types from '@iroha/core/data-model'
+ * import { assertEquals } from '@std/assert/equals'
+ *
+ * const store = types.AssetType.Store
+ * const numeric = types.AssetType.Numeric({ scale: 5 })
+ *
+ * assertEquals(store, { kind: "Store" })
+ * assertEquals(numeric, { kind: "Numeric", value: { scale: 5 } })
+ * ```
+ *
+ * Constructors approach is especially useful when it comes to enums nested into each other, e.g. {@link Parameter}:
+ *
+ * ```ts
+ * import * as types from '@iroha/core/data-model'
+ * import { assertEquals } from '@std/assert/equals'
+ *
+ * const value = types.Parameter.Sumeragi.BlockTime(
+ *   types.Duration.fromMillis(500)
+ * )
+ *
+ * assertEquals(value, {
+ *   kind: "Sumeragi",
+ *   value: {
+ *     kind: "BlockTime",
+ *     value: types.Duration.fromMillis(500)
+ *   }
+ * })
+ * ```
  *
  * @module
  */
@@ -10,3 +76,4 @@ export * from './primitives.ts'
 export * from './compound.ts'
 export * from './_generated_.ts'
 export { Hash, KeyPair, PrivateKey, PublicKey, Signature } from '../crypto/mod.ts'
+export type { Variant, VariantUnit } from '../util.ts'
