+++ packages/core/deno.jsonc
@@ -3,6 +3,7 @@
   "version": "0.0.2",
   "exports": {
     ".": "./mod.ts",
+    "./codec": "./codec.ts",
     "./data-model": "./data-model/mod.ts",
     "./data-model/schema": "./data-model/schema/mod.ts",
     "./data-model/schema-json": "./data-model/schema/json.ts",
